Add files via upload
This machine learning model clusters a set of costumers from its personal data. This prediction is one of 3 clusters. The dataset is taken from the Dry Bean Dataset. So here are the key features of this project:

Prediction is based on these beans attributes

Area (A): The area of a bean zone and the number of pixels within its boundaries.
Perimeter (P): Bean circumference is defined as the length of its border.
Major axis length (L): The distance between the ends of the longest line that can be drawn from a bean.
Minor axis length (l): The longest line that can be drawn from the bean while standing perpendicular to the main axis.
Aspect ratio (K): Defines the relationship between L and l.
Eccentricity (Ec): Eccentricity of the ellipse having the same moments as the region.
Convex area (C): Number of pixels in the smallest convex polygon that can contain the area of a bean seed.
Equivalent diameter (Ed): The diameter of a circle having the same area as a bean seed area.
Extent (Ex): The ratio of the pixels in the bounding box to the bean area.
Solidity (S): Also known as convexity. The ratio of the pixels in the convex shell to those found in beans.
Roundness (R): Calculated with the following formula: (4piA)/(P^2)
Compactness (CO): Measures the roundness of an object: Ed/L
ShapeFactor1 (SF1)
ShapeFactor2 (SF2)
ShapeFactor3 (SF3)
ShapeFactor4 (SF4)
Class (Seker, Barbunya, Bombay, Cali, Dermosan, Horoz and Sira)
The target feature is Class, which tells us which class does the bean belong to.

Data balancing with SMOTE.

Based on Scikit-Learn modules and functions such like:

model_selection.GridSearchCV: Dimensionality reduction.
linear_model.LogisticRegression : Machine model classifier
We've gotten a 91.9% of f1_score